Gate define, middleware design and role wise authorization

+++ app/Http/Controllers/Backend/ModuleController.php
@@ -8,6 +8,7 @@ use App\Models\Module;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Brian2694\Toastr\Facades\Toastr;
+use Illuminate\Support\Facades\Gate;
 
 class ModuleController extends Controller
 {
@@ -16,6 +17,7 @@ class ModuleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('index-module'); //authorize this user to access/give access to admin dashboard;
         $modules = Module::select(['id', 'module_name', 'module_slug', 'updated_at'])->latest()->get();
         return view('admin.pages.module.index', compact('modules'));
     }
@@ -25,6 +27,7 @@ class ModuleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create-module'); //authorize this user to access/give access to admin dashboard;
         return view('admin.pages.module.create');
     }
 
@@ -33,6 +36,7 @@ class ModuleController extends Controller
      */
     public function store(ModuleStoreRequest $request)
     {
+        Gate::authorize('create-module'); //authorize this user to access/give access to admin dashboard;
         Module::updateOrCreate([
             'module_name' => $request->module_name,
             'module_slug' => Str::slug($request->module_name),
@@ -55,6 +59,7 @@ class ModuleController extends Controller
      */
     public function edit(string $id)
     {
+        Gate::authorize('edit-module'); //authorize this user to access/give access to admin dashboard;
         $module = Module::find($id);
         return view('admin.pages.module.edit', compact('module'));
     }
@@ -64,6 +69,7 @@ class ModuleController extends Controller
      */
     public function update(ModuleStoreRequest $request, string $id)
     {
+        Gate::authorize('edit-module'); //authorize this user to access/give access to admin dashboard;
         $module = Module::find($id);
         $module->update([
             'module_name' => $request->module_name,
@@ -79,6 +85,7 @@ class ModuleController extends Controller
      */
     public function destroy(string $id)
     {
+        Gate::authorize('delete-module'); //authorize this user to access/give access to admin dashboard;
         $module = Module::find($id);
         $module->delete();
