+++ app/Http/Controllers/Backend/PermissionController.php
@@ -9,7 +9,7 @@ use App\Models\Permission;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Brian2694\Toastr\Facades\Toastr;
-
+use Illuminate\Support\Facades\Gate;
 
 class PermissionController extends Controller
 {
@@ -18,8 +18,8 @@ class PermissionController extends Controller
      */
     public function index()
     {
+        Gate::authorize('index-permission'); //authorize this user to access/give access to admin dashboard;
         $permissions = Permission::with(['module:id,module_name,module_slug'])->select(['id', 'module_id', 'permission_name', 'permission_slug', 'updated_at'])->latest()->paginate();
-        // return $permissions;
         return view('admin.pages.permission.index', compact('permissions'));
     }
 
@@ -28,6 +28,7 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create-permission'); //authorize this user to access/give access to admin dashboard;
         $modules = Module::select(['id', 'module_name'])->get();
         return view('admin.pages.permission.create', compact('modules'));
     }
@@ -37,6 +38,7 @@ class PermissionController extends Controller
      */
     public function store(PermissionStoreRequest $request)
     {
+        Gate::authorize('create-permission'); //authorize this user to access/give access to admin dashboard;
         Permission::updateOrCreate([
             'module_id' => $request->module_id,
             'permission_name' => $request->permission_name,
@@ -60,6 +62,7 @@ class PermissionController extends Controller
      */
     public function edit(string $id)
     {
+        Gate::authorize('edit-permission'); //authorize this user to access/give access to admin dashboard;
         $permission = Permission::find($id);
         $modules = Module::select(['id', 'module_name'])->get();
         return view('admin.pages.permission.edit', compact('modules', 'permission'));
@@ -70,6 +73,7 @@ class PermissionController extends Controller
      */
     public function update(PermissionStoreRequest $request, string $id)
     {
+        Gate::authorize('edit-permission'); //authorize this user to access/give access to admin dashboard;
         $permission = Permission::find($id);
         $permission->update([
             'module_id' => $request->module_id,
@@ -86,6 +90,7 @@ class PermissionController extends Controller
      */
     public function destroy(string $id)
     {
+        Gate::authorize('delete-permission'); //authorize this user to access/give access to admin dashboard;
         $permission = Permission::find($id);
         $permission->delete();
